Split runtime into multiple files
